Ticket #21430
The Newsletters and press releases are still not in order according to their dates submitted on the back end (the newest should be at the top of the page)
It is currently the opposite.
The search feature on the experience page still isn’t working. It appears that the set up is correct however it's not actually pulling any searches. This goes for the experience landing page and the individual practice experience pages.

>> Done